Update chat room UI and online status handling
Updated the chat room UI in templates/chat.html by introducing better styles and improved the way online status of users is handled. Now, in case of 'User.DoesNotExists' exceptions, the system can handle it more efficiently in app_user/views.py. This will enhance user experience and improve the robustness of chat feature.

For better notifications, a system has been added to broadcast messages when a user joins or leaves the chat. This was implemented in chat_app/consumer.py.

The setting for Redis in chat_app/settings.py was also updated to use variables for host and port for easier configuration.

This all cumulatively leads to an enhanced chat room experience and aids in easier application configuration.