README file
Added a new SQLite data source to the .idea/dataSources.xml to connect to the local database 'db [2]'. This provides an easier way to interact with the local database during development.

Also created a detailed README.md file to explain the key features and overview of the application. This increased the clarity & understanding of the project for new developers and stakeholders.